Extract clinical_notes tests
